+++ database/migrations/2025_07_29_040000_create_career_communications_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->foreignId('contact_id', Contact::class)->nullable()->index();
             $table->string('subject');
             $table->text('body');
-            $table->tinyInteger('sequence')->default(0);
+            $table->integer('sequence')->default(0);
             $table->tinyInteger('public')->default(1);
             $table->tinyInteger('disabled')->default(0);
             $table->timestamps();
