+++ database/migrations/2025_07_29_045000_create_career_notes_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('application_id', Application::class)->nullable();
             $table->string('subject');
             $table->text('body');
-            $table->tinyInteger('sequence')->default(0);
+            $table->integer('sequence')->default(0);
             $table->tinyInteger('public')->default(1);
             $table->tinyInteger('disabled')->default(0);
             $table->timestamps();
